Adding blocking and challenges is next
However we can add all the other methods, with these placement holders for block and challege methods, So we should be able to implement the entire game this way. As such, i think it will be easier to just call challenge and block inside our functions instead doing some stack thing.