We can start looking at using model
The backend isnt even close to done, however the user python program should give a descriptive answer to what it will roughly look like. The confusing part to me is how the ml will actually choose a path with given inputs but its probably not that bad. Once i get the bindings finished for each function, and the game logic done then we start training.